Correct App structure ... something went wrong there xD Fixes #152
